operator installation: maxUnavailable=100% in case of a single replica
The rolling update of the Cilium Operator deployment gets stuck on
single node k8s clusters if the default helm chart values with
readinessprobe, rolling update with maxUnavailable=50% and
podAntiAffinity (which disallows operator pods on the same node) are
applied. The newly createed Pod can't get scheduled before the old
replcia gets terminated.

This change configures the Cilium Operators rolling update to use
maxUnavailable=100% if only one replica is configured and no
userprovided update strategy is provided. This leads to an immediate
termination of the existing Pod on a rolling update.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>